Add a website_leaflet_map addon
This addon can be used as a replacement for website_google_map.
It uses OpenStreetMap tiles and Leaflet.js to add the markers.

The code is loosely modeled after website_google_map.

Unlike the google_map addon though, it doesn't try to geolocalize
partners when coordinates are invalid.

But it tries to zoom the map correctly to fit all the markers.